requirements.txt
Django==5.2.1
matplotlib==3.10.3
numpy==1.23.5
sympy==1.14.0
python-dotenv==1.1.1
openai==0.27.0
djangorestframework==3.16.0